Fix bug where modules do not complete
This works by passing an additional parameter when calling cm#complete.
This parameter prevents caching so the long-term preference of this
parameter is dubious. It may be okay, if completist does it's own
caching, therefore minimizing the performance impact.

A long term solution that seems to be exposing itself is to
auto-complete `` (blank) unless there's a `/` in the typed context, in
which case autocomplete up to the `/`, and nothing after. This will
allow the caching in NCM to work.